training/pyproject.toml
[project]
dependencies = ["torch>=2.6.0", "pandas>=2.2.3", "pillow>=11.2.1", "torchvision>=0.21.0", "numpy>=2.2.4", "matplotlib>=3.10.1", "opencv-python>=4.11.0.86", "torcheval>=0.0.7", "tqdm>=4.67.1"]
requires-python = "==3.12.*"
